add focus prevention for ArcSceneView
Currently if the scene is focussed by the keyboard it causes the scene to break its layout (it recentres in an odd positionI have added some changes to now prevent it from ever being focusable as it is not interactive.

The overide has to be quite manual because the Esri web component manages a lot of its own statemangement and is difficuly to hide from focus.

- Implement useEffect to prevent focus on ArcSceneView and its children.
- Adjust border styles in the globe component to use new bas style guide variables.