Update map and globe components to better track when layers are ready.
Changes were made to the initialisation logic so that it is now possible to check when the initial layers have properly finished updating. This is required because the mapview will signal it is ready when it has finished making network requests which is before the layers and content is actually drawn/rendered in the user interface.

The new logic now waits for both the mapview to be ready, and the initial layers to have stopped updating.

Note that the graticule logic has been moved to a map initialisation command. This allows it to be picked up in these initial layer update checks without any additional hooks being required.

Changes were made to the corresponding test utils such that they now look for the data-ready attribute on the scene/map during the initial load. My expectation is that the tests will be more stable now.